changed block list function to pass in block list, added tests, renamed util to utils module
